spikeagent_0.102.3_gpu/requirements-docker.txt
absl-py==2.2.2
accelerate==1.10.0
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
altair==4.2.2
anndata==0.12.2
annexremote==1.6.6
annotated-types==0.7.0
anthropic==0.49.0
array-api-compat==1.12.0
array_record==0.7.1
asciitree==0.3.3
astunparse==1.6.3
audio-recorder-streamlit==0.0.10
babe==0.0.7
backports.tarfile==1.2.0
beartype==0.14.1
better-abc==0.0.3
blinker==1.9.0
blosc2==3.4.0
bokeh==3.6.3
boto3==1.39.4
botocore==1.39.4
cachetools==5.5.2
chardet==5.2.0
chex==0.1.89
ci-info==0.3.0
click==8.1.8
cloudpickle==3.1.1
config2py==0.1.42
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
crc32c==2.7.1
cryptography==45.0.5
cycler==0.12.1
Cython==3.1.2
dash==3.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2025.3.0
dataclasses-json==0.6.7
datalad==1.2.1
datasets==3.5.0
Deprecated==1.2.18
dill==0.3.8
distrax==0.1.5
distro==1.9.0
dm-tree==0.1.9
docstring_parser==0.16
dol==0.3.30
donfig==0.8.1.post1
dotenv==0.9.9
editdistance==0.8.1
einops==0.8.1
elephant==1.1.1
et_xmlfile==2.0.0
etelemetry==0.3.1
etils==1.12.2
faiss-gpu==1.10.0
fancy-einsum==0.0.3
fasteners==0.19
figurl==0.3.0a1
filelock==3.17.0
filetype==1.2.0
find_libpython==0.4.1
Flask==3.0.3
flatbuffers==25.2.10
flax==0.10.5
fonttools==4.56.0
frontend==0.0.3
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.38.0
google-pasta==0.2.0
google_search_results==2.4.2
googleapis-common-protos==1.69.2
graze==0.1.31
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h5py==3.13.0
harmonypy==0.0.10
hdbscan==0.8.40
hdmf==4.1.0
herdingspikes==0.4.6
httplib2==0.22.0
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanize==4.12.2
i2==0.1.58
igraph==0.11.9
imageio==2.37.0
immutabledict==4.2.1
iso8601==2.1.0
isodate==0.6.1
isosplit5==0.2.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jax==0.6.0
jaxlib==0.6.0
jaxtyping==0.3.2
jeepney==0.9.0
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
kachery==2.0.2
keras==3.9.2
keyring==25.6.0
keyrings.alt==5.0.2
kilosort==4.0.21
kiwisolver==1.4.8
langchain==0.3.20
langchain-anthropic==0.3.10
langchain-community==0.3.19
langchain-core==0.3.67
langchain-experimental==0.3.4
langchain-google-genai==2.1.1
langchain-openai==0.3.11
langchain-text-splitters==0.3.6
langgraph==0.3.8
langgraph-checkpoint==2.0.19
langgraph-prebuilt==0.1.2
langgraph-sdk==0.1.56
lazy-ops==0.2.0
lazy_loader==0.4
legacy-api-wrap==1.4.1
leidenalg==0.10.2
LFPy==2.3.5
LFPykit==0.5.1
libclang==18.1.1
llvmlite==0.42.0
locket==1.0.0
looseversion==1.3.0
lxml==5.3.1
Markdown==3.8
markdown-it-py==3.0.0
marshmallow==3.26.1
matplotlib==3.10.1
mdurl==0.1.2
MEArec==1.9.2
MEAutility==1.5.3
ml_collections==1.1.0
ml_dtypes==0.5.1
more-itertools==10.7.0
mountainsort4==1.0.6
mountainsort5
mpmath==1.3.0
msgpack==1.1.0
mtscomp==1.0.2
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.30.0
natsort==8.4.0
ndindex==1.10.0
neo==0.14.1
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.2
numba==0.59.0
numcodecs==0.15.1
numexpr==2.11.0
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.68.2
openpyxl==3.1.5
opt_einsum==3.4.0
optax==0.2.4
optree==0.15.0
orbax-checkpoint==0.11.12
orjson==3.10.15
pandas==2.2.3
partd==1.4.2
pathlib==1.0.1
patool==4.0.1
patsy==1.0.1
pdoc==15.0.4
pillow==11.1.0
plotly==6.0.1
plotly-express==0.4.1
pooch==1.8.2
probeinterface==0.2.26
promise==2.3
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
prov==2.0.1
puremagic==1.28
py-cpuinfo==9.0.0
py2store==0.1.21
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
PyMuPDF==1.25.5
pynndescent==0.5.13
pynwb==3.0.0
pyparsing==3.2.1
python-dotenv==1.0.1
python-gitlab==6.1.0
python-igraph==0.11.9
pytils==0.4.3
pyxnat==1.6.3
quantities==0.16.1
rdflib==6.3.2
regex==2024.11.6
requests-toolbelt==1.0.0
retrying==1.3.4
rich==14.0.0
rsa==4.9
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.13.0
sae-lens==6.16.0
safetensors==0.5.3
scanpy==1.11.4
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.3
sentencepiece==0.2.0
sentry-sdk==2.35.0
serpapi==0.1.5
session-info2==0.2
simple-parsing==0.1.7
simplejson==3.20.1
smmap==5.0.2
sortingview==0.14.1
spikeextractors==0.9.11
spikeinterface==0.102.3
SQLAlchemy==2.0.39
starlette==0.46.1
statsmodels==0.14.5
streamlit==1.43.2
streamlit-audiorec==0.1.3
stringcase==1.2.0
sympy==1.14.0
tables==3.10.2
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-gan==2.1.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.1
tensorflow-probability==0.25.0
tensorflow-text==2.19.0
tensorstore==0.1.73
termcolor==3.0.1
texttable==1.7.0
tf_keras==2.19.0
tfds-nightly==4.9.8.dev202504200045
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tools==0.1.9
toolz==1.0.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
traits==7.0.2
transformer-lens==2.16.1
transformers==4.51.3
transformers-stream-generator==0.0.5
treescope==0.1.9
typeguard==4.4.4
typing-inspect==0.9.0
typing_extensions==4.14.1
tzdata==2025.1
umap==0.1.1
umap-learn==0.5.7
uvicorn==0.34.0
wadler_lindig==0.1.7
wandb==0.21.1
watchdog==6.0.0
Werkzeug==3.0.6
wrapt==1.17.2
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.18.3
zarr==2.18.7
zstandard==0.23.0
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
attrs
babel
beautifulsoup4
bleach
Brotli
cached-property
certifi
cffi
charset-normalizer
comm
debugpy
decorator
defusedxml
entrypoints
exceptiongroup
executing
fastjsonschema
fqdn
h11
h2
hpack
httpcore
httpx
hyperframe
idna
importlib_metadata
importlib_resources
ipykernel
ipython
ipywidgets
isoduration
jedi
Jinja2
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kaleido
MarkupSafe
matplotlib-inline
mistune
nbclient
nbconvert
nbformat
nest_asyncio
notebook
notebook_shim
overrides
packaging
pandocfilters
parso
pexpect
pickleshare
pkgutil_resolve_name
platformdirs
prometheus_client
prompt_toolkit
psutil
ptyprocess
pure_eval
pycparser
Pygments
PySocks
python-dateutil
python-json-logger
pytz
PyYAML
pyzmq
referencing
requests
rfc3339_validator
rfc3986-validator
rpds-py
Send2Trash
six
sniffio
soupsieve
stack_data
terminado
tinycss2
tomli
tornado
traitlets
types-python-dateutil
typing_utils
uri-template
urllib3
wcwidth
webcolors
webencodings
websocket-client
widgetsnbextension
zipp
